+++ internal/gio/copy.go
@@ -4,19 +4,45 @@ import (
 	"fmt"
 	"io"
 	"os"
+	"path/filepath"
+	"strings"
+
+	"github.com/apex/log"
 )
 
-// CopyFileWithSrcMode copies src into dst with src's file mode.
-func CopyFileWithSrcMode(src, dst string) error {
-	st, err := os.Stat(src)
-	if err != nil {
-		return fmt.Errorf("failed to stat %s: %w", src, err)
-	}
-	return CopyFile(src, dst, st.Mode())
+// Copy recursively copies src into dst with src's file modes.
+func Copy(src, dst string) error {
+	return CopyWithMode(src, dst, 0)
 }
 
-// CopyFile copies src into dst with the given mode.
-func CopyFile(src, dst string, mode os.FileMode) error {
+// CopyWithMode recursively copies src into dst with the given mode.
+// The given mode applies only to files. Their parent dirs will have the same mode as their src counterparts.
+func CopyWithMode(src, dst string, mode os.FileMode) error {
+	return filepath.Walk(src, func(path string, info os.FileInfo, err error) error {
+		if err != nil {
+			return fmt.Errorf("failed to copy %s to %s", src, dst)
+		}
+		// We have the following:
+		// - src = "a/b"
+		// - dst = "dist/linuxamd64/b"
+		// - path = "a/b/c.txt"
+		// So we join "a/b" with "c.txt" and use it as the destination.
+		dst := filepath.Join(dst, strings.Replace(path, src, "", 1))
+		log.WithFields(log.Fields{
+			"src": path,
+			"dst": dst,
+		}).Debug("copying file")
+		if info.IsDir() {
+			return os.MkdirAll(dst, info.Mode())
+		}
+		if mode != 0 {
+			return copyFile(path, dst, mode)
+		}
+		return copyFile(path, dst, info.Mode())
+	})
+}
+
+func copyFile(src, dst string, mode os.FileMode) error {
 	original, err := os.Open(src)
 	if err != nil {
 		return fmt.Errorf("failed to open '%s': %w", src, err)
