+++ internal/pipe/docker/docker.go
@@ -152,19 +152,19 @@ func process(ctx *context.Context, docker config.Docker, artifacts []*artifact.A
 	log := log.WithField("image", images[0])
 	log.Debug("tempdir: " + tmp)
 
-	if err := gio.CopyFileWithSrcMode(docker.Dockerfile, filepath.Join(tmp, "Dockerfile")); err != nil {
+	if err := gio.Copy(docker.Dockerfile, filepath.Join(tmp, "Dockerfile")); err != nil {
 		return fmt.Errorf("failed to link dockerfile: %w", err)
 	}
 	for _, file := range docker.Files {
 		if err := os.MkdirAll(filepath.Join(tmp, filepath.Dir(file)), 0o755); err != nil {
 			return fmt.Errorf("failed to link extra file '%s': %w", file, err)
 		}
-		if err := copyDir(file, filepath.Join(tmp, file)); err != nil {
+		if err := gio.Copy(file, filepath.Join(tmp, file)); err != nil {
 			return fmt.Errorf("failed to link extra file '%s': %w", file, err)
 		}
 	}
 	for _, art := range artifacts {
-		if err := gio.CopyFileWithSrcMode(art.Path, filepath.Join(tmp, filepath.Base(art.Path))); err != nil {
+		if err := gio.Copy(art.Path, filepath.Join(tmp, filepath.Base(art.Path))); err != nil {
 			return fmt.Errorf("failed to link artifact: %w", err)
 		}
 	}
@@ -237,30 +237,6 @@ func processBuildFlagTemplates(ctx *context.Context, docker config.Docker) ([]st
 		buildFlags = append(buildFlags, buildFlag)
 	}
 	return buildFlags, nil
-}
-
-// walks the src, recreating dirs and copying files.
-// TODO: this is very similar to the the same function in the snapcraft package.
-func copyDir(src, dest string) error {
-	return filepath.Walk(src, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-		// We have the following:
-		// - src = "a/b"
-		// - dest = "dist/linuxamd64/b"
-		// - path = "a/b/c.txt"
-		// So we join "a/b" with "c.txt" and use it as the destination.
-		dst := filepath.Join(dest, strings.Replace(path, src, "", 1))
-		log.WithFields(log.Fields{
-			"src": path,
-			"dst": dst,
-		}).Debug("extra file")
-		if info.IsDir() {
-			return os.MkdirAll(dst, info.Mode())
-		}
-		return gio.CopyFileWithSrcMode(path, dst)
-	})
 }
 
 func dockerPush(ctx *context.Context, image *artifact.Artifact) error {
